raspi-cloud-controller/requirements.txt
flask==1.0.2
requests==2.21.0
paho-mqtt==1.4.0